reactApp/my_api_test1/requirements.txt
django
djangorestframework
django_filter
django-cors-headers
